requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.7.0,<3.0.0
langchain>=0.3.21,<1.0.0
openai>=1.68.2,<2.0.0
redis==5.0.1
pypdf2==3.0.1
python-jose==3.3.0
email-validator==2.1.0.post1
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
tiktoken>=0.7,<1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
jsonschema>=4.20.0
langsmith>=0.1.125,<0.4 